content/html/content/test/test_bug596511.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=596511
-->
<head>
  <title>Test for Bug 596511</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=596511">Mozilla Bug 596511</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 596511 **/

function checkRequiredAttribute(element)
{
  ok('required' in element, "select should have a required attribute");

  ok(!element.required, "select required attribute should be disabled");
  is(element.getAttribute('required'), null,
    "select required attribute should be disabled");

  element.required = true;
  ok(element.required, "select required attribute should be enabled");
  isnot(element.getAttribute('required'), null,
    "select required attribute should be enabled");

  element.removeAttribute('required');
  element.setAttribute('required', '');
  ok(element.required, "select required attribute should be enabled");
  isnot(element.getAttribute('required'), null,
    "select required attribute should be enabled");

  element.removeAttribute('required');
  ok(!element.required, "select required attribute should be disabled");
  is(element.getAttribute('required'), null,
    "select required attribute should be disabled");
}

function checkRequiredAndOptionalSelectors(element)
{
  is(document.querySelector("select:optional"), element, "select should be optional");
  is(document.querySelector("select:required"), null, "select shouldn't be required");

  element.required = true;

  is(document.querySelector("select:optional"), null, "select shouldn't be optional");
  is(document.querySelector("select:required"), element, "select should be required");

  element.required = false;
}

var select = document.createElement("select");
var content = document.getElementById('content');
content.appendChild(select);

checkRequiredAttribute(select);
checkRequiredAndOptionalSelectors(select);

</script>
</pre>
</body>
</html>
